+++ src/coq_elpi_utils.mli
@@ -19,7 +19,10 @@ val mk_gforall : Glob_term.glob_constr -> Glob_term.glob_decl list -> Glob_term.
 val mk_gfun : Glob_term.glob_constr -> Glob_term.glob_decl list -> Glob_term.glob_constr
 val manual_implicit_of_binding_kind : Names.Name.t -> Glob_term.binding_kind -> (Names.Name.t * bool) option CAst.t
 val manual_implicit_of_gdecl : Glob_term.glob_decl -> (Names.Name.t * bool) option CAst.t
+
 val lookup_inductive : Environ.env -> Names.inductive -> Declarations.mutual_inductive_body * Declarations.one_inductive_body
+val locate_gref : string -> Names.GlobRef.t
+val locate_qualid : Libnames.qualid -> [ `Gref of Names.GlobRef.t | `Abbrev of Globnames.syndef_name ] option
 
 val fold_elpi_term :
   (depth:int -> 'a -> Elpi.API.Data.term -> 'a) ->